Create layout & top navigation bar and side bar
In order to accelerate the development, we implement bootstrap to our project. In
addition, our design employs Paper Dashboard Layout, which means that the entry point
will always from the Dashboard.vue page, so DashboardLayout.vue just only for router.
The Plugins and assets include all the needs for component and design.